+++ diary/src/main/java/com/teamauc/diary/controller/ReplyController.java
@@ -5,7 +5,10 @@ import com.teamauc.diary.domain.Reply;
 import com.teamauc.diary.domain.User;
 import com.teamauc.diary.dto.MessageResponseDto;
 import com.teamauc.diary.dto.ResultDto;
+import com.teamauc.diary.exception.InvalidApproachException;
+import com.teamauc.diary.exception.UnauthorizedException;
 import com.teamauc.diary.service.DiaryService;
+import com.teamauc.diary.service.JwtService;
 import com.teamauc.diary.service.ReplyService;
 import com.teamauc.diary.service.UserService;
 import lombok.*;
@@ -26,13 +29,22 @@ public class ReplyController {
     private final UserService userService;
     private final DiaryService diaryService;
     private final ReplyService replyService;
+    private final JwtService jwtService;
 
 
     @PostMapping
     public CreateReplyResponseDto createReply (@RequestBody CreateReplyRequestDto request) {
 
+        if (!jwtService.isValidUser())
+            throw new InvalidApproachException("사용자 인증 실패");
+
+        String currentUid = jwtService.getUserId();
         Diary diary = diaryService.findByDiaryId(request.getDiaryId());
-        User user = userService.SearchUserById(request.getUid());
+        User user = userService.SearchUserById(jwtService.getUserId());
+
+        boolean isMyDiary = currentUid.equals(diary.getUser().getUid());
+
+        if(!isMyDiary&&diary.isSecret()) throw new UnauthorizedException("남의 비밀일기에 댓글을 달 수 없습니다.");
 
         Reply reply = Reply.createReply(diary,user,LocalDateTime.now(),request.isSecret(), request.getContent());
 
@@ -44,6 +56,15 @@ public class ReplyController {
 
     @GetMapping("/user/{uid}")
     public ResultDto readReplyByUserId (@PathVariable ("uid") String uid){
+
+        if (!jwtService.isValidUser())
+            throw new InvalidApproachException("사용자 인증 실패");
+
+        String currentUid = jwtService.getUserId();
+
+        boolean isMine = currentUid.equals(uid);
+
+        if(!isMine) throw new UnauthorizedException();
 
         List<ReadReplyResponseDto> replies = replyService.findByUserId(uid)
                 .stream()
@@ -57,9 +78,35 @@ public class ReplyController {
     @GetMapping("/diary/{id}")
     public ResultDto readReplyByDiaryId (@PathVariable ("id") String id){
 
+        if (!jwtService.isValidUser())
+            throw new InvalidApproachException("사용자 인증 실패");
+
+        String currentUid = jwtService.getUserId();
+        Diary diary = diaryService.findByDiaryId(id);
+        String diaryOwnerUid = diary.getUser().getUid();
+
+        boolean isMine = diaryOwnerUid.equals(currentUid);
+
+        if(diary.isSecret()&&!isMine) throw new UnauthorizedException("비밀 일기에 달린 댓글입니다.");
+
         List<ReadReplyResponseDto> replies = replyService.findByDiaryId(id)
                 .stream()
-                .map(reply -> new ReadReplyResponseDto(reply))
+                .map(reply -> {
+
+                    if (!reply.isSecret()) return new ReadReplyResponseDto(reply);
+
+                    boolean isQualifiedUser = (reply.getId().equals(currentUid))||isMine;
+
+                    if (isQualifiedUser) {
+                        return new ReadReplyResponseDto(reply);
+                    }
+                    else {
+                        reply.setContent("비밀 댓글입니다.");
+                        return new ReadReplyResponseDto(reply);
+                    }
+
+
+                })
                 .collect(Collectors.toList());
 
         return new ResultDto(replies);
@@ -69,7 +116,16 @@ public class ReplyController {
     @GetMapping("/{id}")
     public ResultDto readReplyById(@PathVariable ("id") Long id){
 
+        if(!jwtService.isValidUser())
+            throw new InvalidApproachException("사용자 인증 실패");
+
+        String currentUid = jwtService.getUserId();
         Reply reply = replyService.findById(id);
+
+        boolean isMine = currentUid.equals(reply.getUser().getUid());
+
+        if(!isMine) throw new UnauthorizedException();
+
 
         ReadReplyResponseDto replydto = new ReadReplyResponseDto(reply);
 
@@ -79,6 +135,16 @@ public class ReplyController {
     @PutMapping("/{id}")
     public MessageResponseDto updateReply(@PathVariable ("id") Long id,@RequestBody UpdateReplyRequestDto request){
 
+        if(!jwtService.isValidUser())
+            throw new InvalidApproachException("사용자 인증 실패");
+
+        String currentUid = jwtService.getUserId();
+        Reply reply = replyService.findById(id);
+
+        boolean isMine = currentUid.equals(reply.getUser().getUid());
+
+        if(!isMine) throw new UnauthorizedException();
+
         replyService.update(id,request.isSecret(), request.getContent());
 
         return new MessageResponseDto("댓글 수정 완료.");
@@ -86,6 +152,16 @@ public class ReplyController {
 
     @DeleteMapping("/{id}")
     public MessageResponseDto deleteReply(@PathVariable ("id") Long id){
+
+        if(!jwtService.isValidUser())
+            throw new InvalidApproachException("사용자 인증 실패");
+
+        String currentUid = jwtService.getUserId();
+        Reply reply = replyService.findById(id);
+
+        boolean isMine = currentUid.equals(reply.getUser().getUid());
+
+        if(!isMine) throw new UnauthorizedException();
 
         replyService.delete(id);
 
@@ -98,8 +174,6 @@ public class ReplyController {
     static class CreateReplyRequestDto{
 
         private String diaryId;
-
-        private String uid;
 
         private boolean secret;
